Find correct virtualenv site-packages.
This patch, taken from https://github.com/davidhalter/jedi/pull/829/commits/2ca6dd4a98a9f420d5c547c08aa1f9dfd6bd9801
does not try to guess the python version name used in the virtualenv path.
The previous code looks up the python version from the running python instance,
which in the case of jedi will be the system python, and not the virtualenv
python. This would break completion in cases where the system python is
python 3.5.2 and the virtualenv is running on 3.4.3 or 3.6 (using pyenv).